@HeroicKatora 2025-05-11
The data from my laptop (Intel i7-8565U)  is not nearly as convincing but still we get a few percent:

```
msb-8/benches/Cargo-8-msb.lzw/8687
                        time:   [224.64 µs 224.85 µs 225.17 µs]
                        thrpt:  [111.15 MiB/s 111.31 MiB/s 111.41 MiB/s]
                 change:
                        time:   [-6.7571% -5.4585% -4.3846%] (p = 0.00 < 0.05)
                        thrpt:  [+4.5857% +5.7737% +7.2468%]
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) high mild
  7 (7.00%) high severe

msb-8/benches/binary-8-msb.lzw/1428141
                        time:   [38.252 ms 38.568 ms 38.936 ms]
                        thrpt:  [74.887 MiB/s 75.602 MiB/s 76.226 MiB/s]
                 change:
                        time:   [+0.0035% +0.7845% +1.6919%] (p = 0.09 > 0.05)
                        thrpt:  [-1.6638% -0.7784% -0.0035%]
                        No change in performance detected.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

msb-8/benches/lib-8-msb.lzw/731
                        time:   [14.825 µs 14.839 µs 14.856 µs]
                        thrpt:  [71.832 MiB/s 71.918 MiB/s 71.985 MiB/s]
                 change:
                        time:   [-5.5125% -4.3853% -3.3196%] (p = 0.00 < 0.05)
                        thrpt:  [+3.4336% +4.5865% +5.8341%]
                        Performance has improved.
Found 21 outliers among 100 measurements (21.00%)
  8 (8.00%) high mild
  13 (13.00%) high severe
```