@WorldSEnder 2025-05-12
Getting very good numbers here apart from desktop systems being noisy as always.

CPU: i7-13700K

```
msb-8/benches/Cargo-8-msb.lzw/8687                                                                             
                        time:   [37.299 µs 37.334 µs 37.369 µs]
                        thrpt:  [669.77 MiB/s 670.38 MiB/s 671.02 MiB/s]
                 change:
                        time:   [-23.984% -23.793% -23.637%] (p = 0.00 < 0.05)
                        thrpt:  [+30.953% +31.221% +31.551%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe

msb-8/benches/binary-8-msb.lzw/1428141                                                                            
                        time:   [9.6720 ms 9.6927 ms 9.7296 ms]
                        thrpt:  [299.68 MiB/s 300.82 MiB/s 301.47 MiB/s]
                 change:
                        time:   [-7.1821% -6.9612% -6.5683%] (p = 0.00 < 0.05)
                        thrpt:  [+7.0300% +7.4821% +7.7379%]
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

msb-8/benches/lib-8-msb.lzw/731                                                                             
                        time:   [2.5581 µs 2.5641 µs 2.5776 µs]
                        thrpt:  [414.02 MiB/s 416.19 MiB/s 417.16 MiB/s]
                 change:
                        time:   [-14.760% -14.526% -14.072%] (p = 0.00 < 0.05)
                        thrpt:  [+16.377% +16.994% +17.315%]
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  4 (4.00%) high severe
```