@halvko 2025-05-14
I was sitting around with an ARM processor (Apple M3 Pro) to test this on:

Current `main` (620efa59) vs `performance-tuning` (3bdf6844):
```
msb-8/benches/Cargo-8-msb.lzw/8687
                        time:   [35.623 µs 35.671 µs 35.733 µs]
                        thrpt:  [700.43 MiB/s 701.65 MiB/s 702.59 MiB/s]
Found 22 outliers among 100 measurements (22.00%)
  10 (10.00%) low mild
  2 (2.00%) high mild
  10 (10.00%) high severe

msb-8/benches/binary-8-msb.lzw/1428141
                        time:   [9.2212 ms 9.2227 ms 9.2243 ms]
                        thrpt:  [316.10 MiB/s 316.16 MiB/s 316.21 MiB/s]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

msb-8/benches/lib-8-msb.lzw/731
                        time:   [2.7868 µs 2.7897 µs 2.7933 µs]
                        thrpt:  [382.05 MiB/s 382.54 MiB/s 382.93 MiB/s]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
```